Fix: Reponsive page
