--- src/components/features/layout/Header/HamburguerMenu/Styles.js
@@ -1,15 +1,0 @@
-import { Menu } from "antd";
-import styled from "styled-components";
-import { breakpoints } from "../../../../../styles/stylesVariables";
-
-export const Hamburguer = styled(Menu)`
-  width: 0rem;
-  border-bottom: none;
-  background-color: black;
-  @media (max-width: ${breakpoints.smallDevice}) {
-    width: 2rem;
-  }
-  @media (max-width: ${breakpoints.mobile}) {
-    width: 6rem;
-  }
-`;
