+++ src/components/features/layout/Header/Styles.js
@@ -1,5 +1,6 @@
 import styled from "styled-components";
 import { breakpoints } from "../../../../styles/stylesVariables";
+import { Menu } from "antd";
 
 export const Container = styled.div`
   height: 7rem;
@@ -8,15 +9,12 @@ export const Container = styled.div`
   background-color: black;
   align-items: center;
   justify-content: space-evenly;
-  @media (max-width: ${breakpoints.smallDevice}) {
-    justify-content: space-between;
-  }
 `;
 
 export const City = styled.img`
   width: 50px;
   display: flex;
-  margin-left: 4rem;
+  margin-left: 3rem;
   @media (max-width: ${breakpoints.smallDevice}) {
     margin-left: 0;
   }
@@ -29,8 +27,8 @@ export const Column = styled.div`
   padding-left: 15px;
   padding-right: 0;
   border-left: 2px solid white;
-  @media (max-width: ${breakpoints.mobile}) {
-    border-left: none;
+  @media (max-width: ${breakpoints.smallTablet}) {
+    border-left: 0;
   }
 `;
 
@@ -38,7 +36,7 @@ export const SocialMedias = styled.div`
   display: flex;
   gap: 10px;
   justify-content: center;
-  @media (max-width: ${breakpoints.mobile}) {
+  @media (max-width: ${breakpoints.smallTablet}) {
     display: none;
   }
 `;
@@ -54,7 +52,21 @@ export const Button = styled.button`
   margin-left: 20%;
   display: flex;
   align-items: center;
-  @media (max-width: ${breakpoints.mobile}) {
+  padding-left: 0;
+  @media (max-width: ${breakpoints.smallTablet}) {
     width: 110%;
+    margin-left: 0;
+  }
+`;
+
+export const Hamburguer = styled(Menu)`
+  width: 0rem;
+  border-bottom: none;
+  background-color: black;
+  @media (max-width: ${breakpoints.smallTablet}) {
+    width: 6rem;
+  }
+  @media (max-width: ${breakpoints.smallDevice}) {
+    width: 4rem;
   }
 `;
