+++ src/components/index.js
@@ -5,4 +5,7 @@ export { default as Link } from "./common/Link/Link.js";
 export { default as Header } from "./features/layout/Header/Header.jsx";
 export { default as Footer } from "./features/layout/Footer/Footer.jsx";
 export { default as AppLayout } from "./features/layout/AppLayout/AppLayout.jsx";
-export { default as HamburgerMenu } from "./features/layout/Header/HamburguerMenu/HambuguerMenu.jsx";
+export { default as HamburgerMenu } from "./features/layout/Header/HambuguerMenu.jsx";
+export { default as ModalDeleteUser } from "./features/modals/ModalDeleteUser/ModalDeleteUser.jsx";
+export { default as Table } from "./features/Table/Table.jsx";
+export { default as SearchBar } from "./features/SearchBar/SearchBar.jsx";
